% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semidist-package.R
\docType{package}
\name{semidist-package}
\alias{semidist}
\alias{semidist-package}
\title{semidist: Measure Dependence Between Categorical and Continuous Variables}
\description{
Semi-distance and mean-variance (MV) index are proposed to measure the dependence between a categorical random variable and a continuous variable. Test of independence and feature screening for classification problems can be implemented via the two dependence measures. For the details of the methods, see Zhong et al. (2023) \url{https://www.tandfonline.com/doi/full/10.1080/01621459.2023.2284988}; Cui and Zhong (2019) \url{https://www.sciencedirect.com/science/article/abs/pii/S016794731930115X}; Cui, Li and Zhong (2015) \url{https://www.tandfonline.com/doi/full/10.1080/01621459.2014.920256}.
}
\author{
\strong{Maintainer}: Zhuoxi Li \email{chainchei@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Wei Zhong \email{wzhong@xmu.edu.cn}
  \item Wenwen Guo \email{guowenwen114@163.com}
  \item Hengjian Cui \email{hjcui@bnu.edu.cn}
  \item Runze Li \email{ril4@psu.edu}
}

}
\keyword{internal}
