% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tr_estimate}
\alias{tr_estimate}
\title{Estimate the trace of the covariance matrix and its square}
\usage{
tr_estimate(X)
}
\arguments{
\item{X}{The design matrix.}
}
\value{
A list with elements:
\itemize{
\item \code{tr_S}: estimate for trace of \eqn{\Sigma};
\item \code{tr_S2}: estimate for trace of \eqn{\Sigma^2}.
}
}
\description{
For a design matrix \eqn{\mathbf{X}}, estimate the trace of its covariance matrix \eqn{\Sigma = \mathrm{cov}(\mathbf{X})},
and the square of covariance matrix \eqn{\Sigma^2}.
}
